requirements.txt
fastapi
uvicorn
pydantic
pymongo
motor
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
starlette
google-api-python-client
requires.io
python-multipart

